pyproject.toml
[tool.poetry]
name = "kml-16s-2"
version = "0.1.0"
description = "KM DJH 16S 分析流程"
authors = ["Kristoff <kr1stoff7m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
pandas = "^2.2.2"
pytest = "^8.3.2"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "primary"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "supplemental"

[tool.pytest.ini_options]
testpaths = [
#    'tests/test_fastq.py',
    'tests/test_snakemake.py',
]